requirements.txt
APScheduler==3.6.3
beautifulsoup4==4.11.2
cachetools==4.2.2
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.0
currency-symbols==2.0.3
cycler==0.12.1
fonttools==4.49.0
idna==3.6
kiwisolver==1.4.5
matplotlib==3.8.0
memory-profiler==0.58.0
numpy==1.24.1
packaging==23.2
pandas==2.2.1
pillow==10.2.0
psutil==5.9.8
pyparsing==3.1.1
python-dateutil==2.9.0.post0
python-telegram-bot==13.15
pytz==2024.1
requests==2.31.0
schedule==1.2.1
six==1.16.0
soupsieve==2.5
tornado==6.1
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
